Cost of Living Analysis
This project focuses on analyzing the cost of living in various cities and countries worldwide. The dataset includes a wide range of economic indicators, from the price of basic commodities to the cost of housing, transportation, and entertainment. By leveraging Power BI, we aim to gain valuable insights into the economic disparities between regions and understand the factors contributing to the varying costs of living. This project serves as an exercise in data visualization, analysis, and interpretation, providing a comprehensive view of the world's economic landscape.